+++ spec/support/fixtures/person.rb
@@ -1,12 +1,20 @@
 class Person 
-  attr_accessor :country, :country_options
+  attr_accessor :country, :country_options, :provinces, :provinces_options
 
   def initialize   
     self.country_options=["", "Canada", "US", "Mexico"]
     self.country = "Canada"
-  end
-
-  def reset_country
-    self.country = "Canada"
+    self.provinces_options=[
+      "",
+      "Quebec",
+      "Ontario",
+      "Manitoba",
+      "Saskatchewan",
+      "Alberta",
+      "British Columbia",
+      "Nova Skotia",
+      "Newfoundland",
+    ]
+    self.provinces = ["Quebec", "Manitoba", "Alberta"]
   end
 end